Add tweet id and author screen_name to what is available in TweetDetail
Tweet ID should be reliably there, but screen_name may be retrieved
correctly. It is nested inside a user json key and the fields call to
retrieve it may or may not be smart about looking into nested key
structures. I couldn't figure out how to test if the code was doing the
right thing.